Move decParts tests to own file
